[Chore] AppInterceptor 클래스 파일 생성
